% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCI.R
\docType{data}
\name{BCI}
\alias{BCI}
\title{Barro-Colorado Island Tree Counts}
\format{
A data frame with 50 rows and 225 columns:
\describe{
\item{Abarema.macradenia}{Count for Abarema.macradenia}
\item{Vachellia.melanoceras}{Count for Vachellia.melanoceras}
\item{Acalypha.diversifolia}{Count for Acalypha.diversifolia}
\item{Acalypha.macrostachya}{Count for Acalypha.macrostachya}
\item{Adelia.triloba}{Count for Adelia.triloba}
\item{Aegiphila.panamensis}{Count for Aegiphila.panamensis}
\item{Alchornea.costaricensis}{Count for Alchornea.costaricensis}
\item{Alchornea.latifolia}{Count for Alchornea.latifolia}
\item{Alibertia.edulis}{Count for Alibertia.edulis}
\item{Allophylus.psilospermus}{Count for Allophylus.psilospermus}
\item{Alseis.blackiana}{Count for Alseis.blackiana}
\item{Amaioua.corymbosa}{Count for Amaioua.corymbosa}
\item{Anacardium.excelsum}{Count for Anacardium.excelsum}
\item{Andira.inermis}{Count for Andira.inermis}
\item{Annona.spraguei}{Count for Annona.spraguei}
\item{Apeiba.glabra}{Count for Apeiba.glabra}
\item{Apeiba.tibourbou}{Count for Apeiba.tibourbou}
\item{Aspidosperma.desmanthum}{Count for Aspidosperma.desmanthum}
\item{Astrocaryum.standleyanum}{Count for Astrocaryum.standleyanum}
\item{Astronium.graveolens}{Count for Astronium.graveolens}
\item{Attalea.butyracea}{Count for Attalea.butyracea}
\item{Banara.guianensis}{Count for Banara.guianensis}
\item{Beilschmiedia.pendula}{Count for Beilschmiedia.pendula}
\item{Brosimum.alicastrum}{Count for Brosimum.alicastrum}
\item{Brosimum.guianense}{Count for Brosimum.guianense}
\item{Calophyllum.longifolium}{Count for Calophyllum.longifolium}
\item{Casearia.aculeata}{Count for Casearia.aculeata}
\item{Casearia.arborea}{Count for Casearia.arborea}
\item{Casearia.commersoniana}{Count for Casearia.commersoniana}
\item{Casearia.guianensis}{Count for Casearia.guianensis}
\item{Casearia.sylvestris}{Count for Casearia.sylvestris}
\item{Cassipourea.guianensis}{Count for Cassipourea.guianensis}
\item{Cavanillesia.platanifolia}{Count for Cavanillesia.platanifolia}
\item{Cecropia.insignis}{Count for Cecropia.insignis}
\item{Cecropia.obtusifolia}{Count for Cecropia.obtusifolia}
\item{Cedrela.odorata}{Count for Cedrela.odorata}
\item{Ceiba.pentandra}{Count for Ceiba.pentandra}
\item{Celtis.schippii}{Count for Celtis.schippii}
\item{Cespedesia.spathulata}{Count for Cespedesia.spathulata}
\item{Chamguava.schippii}{Count for Chamguava.schippii}
\item{Chimarrhis.parviflora}{Count for Chimarrhis.parviflora}
\item{Maclura.tinctoria}{Count for Maclura.tinctoria}
\item{Chrysochlamys.eclipes}{Count for Chrysochlamys.eclipes}
\item{Chrysophyllum.argenteum}{Count for Chrysophyllum.argenteum}
\item{Chrysophyllum.cainito}{Count for Chrysophyllum.cainito}
\item{Coccoloba.coronata}{Count for Coccoloba.coronata}
\item{Coccoloba.manzinellensis}{Count for Coccoloba.manzinellensis}
\item{Colubrina.glandulosa}{Count for Colubrina.glandulosa}
\item{Cordia.alliodora}{Count for Cordia.alliodora}
\item{Cordia.bicolor}{Count for Cordia.bicolor}
\item{Cordia.lasiocalyx}{Count for Cordia.lasiocalyx}
\item{Coussarea.curvigemma}{Count for Coussarea.curvigemma}
\item{Croton.billbergianus}{Count for Croton.billbergianus}
\item{Cupania.cinerea}{Count for Cupania.cinerea}
\item{Cupania.latifolia}{Count for Cupania.latifolia}
\item{Cupania.rufescens}{Count for Cupania.rufescens}
\item{Cupania.seemannii}{Count for Cupania.seemannii}
\item{Dendropanax.arboreus}{Count for Dendropanax.arboreus}
\item{Desmopsis.panamensis}{Count for Desmopsis.panamensis}
\item{Diospyros.artanthifolia}{Count for Diospyros.artanthifolia}
\item{Dipteryx.oleifera}{Count for Dipteryx.oleifera}
\item{Drypetes.standleyi}{Count for Drypetes.standleyi}
\item{Elaeis.oleifera}{Count for Elaeis.oleifera}
\item{Enterolobium.schomburgkii}{Count for Enterolobium.schomburgkii}
\item{Erythrina.costaricensis}{Count for Erythrina.costaricensis}
\item{Erythroxylum.macrophyllum}{Count for Erythroxylum.macrophyllum}
\item{Eugenia.florida}{Count for Eugenia.florida}
\item{Eugenia.galalonensis}{Count for Eugenia.galalonensis}
\item{Eugenia.nesiotica}{Count for Eugenia.nesiotica}
\item{Eugenia.oerstediana}{Count for Eugenia.oerstediana}
\item{Faramea.occidentalis}{Count for Faramea.occidentalis}
\item{Ficus.colubrinae}{Count for Ficus.colubrinae}
\item{Ficus.costaricana}{Count for Ficus.costaricana}
\item{Ficus.insipida}{Count for Ficus.insipida}
\item{Ficus.maxima}{Count for Ficus.maxima}
\item{Ficus.obtusifolia}{Count for Ficus.obtusifolia}
\item{Ficus.popenoei}{Count for Ficus.popenoei}
\item{Ficus.tonduzii}{Count for Ficus.tonduzii}
\item{Ficus.trigonata}{Count for Ficus.trigonata}
\item{Ficus.yoponensis}{Count for Ficus.yoponensis}
\item{Garcinia.intermedia}{Count for Garcinia.intermedia}
\item{Garcinia.madruno}{Count for Garcinia.madruno}
\item{Genipa.americana}{Count for Genipa.americana}
\item{Guapira.myrtiflora}{Count for Guapira.myrtiflora}
\item{Guarea.fuzzy}{Count for Guarea.fuzzy}
\item{Guarea.grandifolia}{Count for Guarea.grandifolia}
\item{Guarea.guidonia}{Count for Guarea.guidonia}
\item{Guatteria.dumetorum}{Count for Guatteria.dumetorum}
\item{Guazuma.ulmifolia}{Count for Guazuma.ulmifolia}
\item{Guettarda.foliacea}{Count for Guettarda.foliacea}
\item{Gustavia.superba}{Count for Gustavia.superba}
\item{Hampea.appendiculata}{Count for Hampea.appendiculata}
\item{Hasseltia.floribunda}{Count for Hasseltia.floribunda}
\item{Heisteria.acuminata}{Count for Heisteria.acuminata}
\item{Heisteria.concinna}{Count for Heisteria.concinna}
\item{Hirtella.americana}{Count for Hirtella.americana}
\item{Hirtella.triandra}{Count for Hirtella.triandra}
\item{Hura.crepitans}{Count for Hura.crepitans}
\item{Hieronyma.alchorneoides}{Count for Hieronyma.alchorneoides}
\item{Inga.acuminata}{Count for Inga.acuminata}
\item{Inga.cocleensis}{Count for Inga.cocleensis}
\item{Inga.goldmanii}{Count for Inga.goldmanii}
\item{Inga.laurina}{Count for Inga.laurina}
\item{Inga.semialata}{Count for Inga.semialata}
\item{Inga.nobilis}{Count for Inga.nobilis}
\item{Inga.oerstediana}{Count for Inga.oerstediana}
\item{Inga.pezizifera}{Count for Inga.pezizifera}
\item{Inga.punctata}{Count for Inga.punctata}
\item{Inga.ruiziana}{Count for Inga.ruiziana}
\item{Inga.sapindoides}{Count for Inga.sapindoides}
\item{Inga.spectabilis}{Count for Inga.spectabilis}
\item{Inga.umbellifera}{Count for Inga.umbellifera}
\item{Jacaranda.copaia}{Count for Jacaranda.copaia}
\item{Lacistema.aggregatum}{Count for Lacistema.aggregatum}
\item{Lacmellea.panamensis}{Count for Lacmellea.panamensis}
\item{Laetia.procera}{Count for Laetia.procera}
\item{Laetia.thamnia}{Count for Laetia.thamnia}
\item{Lafoensia.punicifolia}{Count for Lafoensia.punicifolia}
\item{Licania.hypoleuca}{Count for Licania.hypoleuca}
\item{Licania.platypus}{Count for Licania.platypus}
\item{Lindackeria.laurina}{Count for Lindackeria.laurina}
\item{Lonchocarpus.heptaphyllus}{Count for Lonchocarpus.heptaphyllus}
\item{Luehea.seemannii}{Count for Luehea.seemannii}
\item{Macrocnemum.roseum}{Count for Macrocnemum.roseum}
\item{Maquira.guianensis.costaricana}{Count for Maquira.guianensis.costaricana}
\item{Margaritaria.nobilis}{Count for Margaritaria.nobilis}
\item{Marila.laxiflora}{Count for Marila.laxiflora}
\item{Maytenus.schippii}{Count for Maytenus.schippii}
\item{Miconia.affinis}{Count for Miconia.affinis}
\item{Miconia.argentea}{Count for Miconia.argentea}
\item{Miconia.elata}{Count for Miconia.elata}
\item{Miconia.hondurensis}{Count for Miconia.hondurensis}
\item{Mosannona.garwoodii}{Count for Mosannona.garwoodii}
\item{Myrcia.gatunensis}{Count for Myrcia.gatunensis}
\item{Myrospermum.frutescens}{Count for Myrospermum.frutescens}
\item{Nectandra.cissiflora}{Count for Nectandra.cissiflora}
\item{Nectandra.lineata}{Count for Nectandra.lineata}
\item{Nectandra.purpurea}{Count for Nectandra.purpurea}
\item{Ochroma.pyramidale}{Count for Ochroma.pyramidale}
\item{Ocotea.cernua}{Count for Ocotea.cernua}
\item{Ocotea.oblonga}{Count for Ocotea.oblonga}
\item{Ocotea.puberula}{Count for Ocotea.puberula}
\item{Ocotea.whitei}{Count for Ocotea.whitei}
\item{Oenocarpus.mapora}{Count for Oenocarpus.mapora}
\item{Ormosia.amazonica}{Count for Ormosia.amazonica}
\item{Ormosia.coccinea}{Count for Ormosia.coccinea}
\item{Ormosia.macrocalyx}{Count for Ormosia.macrocalyx}
\item{Pachira.quinata}{Count for Pachira.quinata}
\item{Pachira.sessilis}{Count for Pachira.sessilis}
\item{Perebea.xanthochyma}{Count for Perebea.xanthochyma}
\item{Cinnamomum.triplinerve}{Count for Cinnamomum.triplinerve}
\item{Picramnia.latifolia}{Count for Picramnia.latifolia}
\item{Piper.reticulatum}{Count for Piper.reticulatum}
\item{Platymiscium.pinnatum}{Count for Platymiscium.pinnatum}
\item{Platypodium.elegans}{Count for Platypodium.elegans}
\item{Posoqueria.latifolia}{Count for Posoqueria.latifolia}
\item{Poulsenia.armata}{Count for Poulsenia.armata}
\item{Pourouma.bicolor}{Count for Pourouma.bicolor}
\item{Pouteria.fossicola}{Count for Pouteria.fossicola}
\item{Pouteria.reticulata}{Count for Pouteria.reticulata}
\item{Pouteria.stipitata}{Count for Pouteria.stipitata}
\item{Prioria.copaifera}{Count for Prioria.copaifera}
\item{Protium.costaricense}{Count for Protium.costaricense}
\item{Protium.panamense}{Count for Protium.panamense}
\item{Protium.tenuifolium}{Count for Protium.tenuifolium}
\item{Pseudobombax.septenatum}{Count for Pseudobombax.septenatum}
\item{Psidium.friedrichsthalianum}{Count for Psidium.friedrichsthalianum}
\item{Psychotria.grandis}{Count for Psychotria.grandis}
\item{Pterocarpus.rohrii}{Count for Pterocarpus.rohrii}
\item{Quararibea.asterolepis}{Count for Quararibea.asterolepis}
\item{Quassia.amara}{Count for Quassia.amara}
\item{Randia.armata}{Count for Randia.armata}
\item{Sapium.broadleaf}{Count for Sapium.broadleaf}
\item{Sapium.glandulosum}{Count for Sapium.glandulosum}
\item{Schizolobium.parahyba}{Count for Schizolobium.parahyba}
\item{Senna.dariensis}{Count for Senna.dariensis}
\item{Simarouba.amara}{Count for Simarouba.amara}
\item{Siparuna.guianensis}{Count for Siparuna.guianensis}
\item{Siparuna.pauciflora}{Count for Siparuna.pauciflora}
\item{Sloanea.terniflora}{Count for Sloanea.terniflora}
\item{Socratea.exorrhiza}{Count for Socratea.exorrhiza}
\item{Solanum.hayesii}{Count for Solanum.hayesii}
\item{Sorocea.affinis}{Count for Sorocea.affinis}
\item{Spachea.membranacea}{Count for Spachea.membranacea}
\item{Spondias.mombin}{Count for Spondias.mombin}
\item{Spondias.radlkoferi}{Count for Spondias.radlkoferi}
\item{Sterculia.apetala}{Count for Sterculia.apetala}
\item{Swartzia.simplex.var.grandiflora}{Count for Swartzia.simplex.var.grandiflora}
\item{Swartzia.simplex.continentalis}{Count for Swartzia.simplex.continentalis}
\item{Symphonia.globulifera}{Count for Symphonia.globulifera}
\item{Handroanthus.guayacan}{Count for Handroanthus.guayacan}
\item{Tabebuia.rosea}{Count for Tabebuia.rosea}
\item{Tabernaemontana.arborea}{Count for Tabernaemontana.arborea}
\item{Tachigali.versicolor}{Count for Tachigali.versicolor}
\item{Talisia.nervosa}{Count for Talisia.nervosa}
\item{Talisia.princeps}{Count for Talisia.princeps}
\item{Terminalia.amazonia}{Count for Terminalia.amazonia}
\item{Terminalia.oblonga}{Count for Terminalia.oblonga}
\item{Tetragastris.panamensis}{Count for Tetragastris.panamensis}
\item{Tetrathylacium.johansenii}{Count for Tetrathylacium.johansenii}
\item{Theobroma.cacao}{Count for Theobroma.cacao}
\item{Thevetia.ahouai}{Count for Thevetia.ahouai}
\item{Tocoyena.pittieri}{Count for Tocoyena.pittieri}
\item{Trattinnickia.aspera}{Count for Trattinnickia.aspera}
\item{Trema.micrantha}{Count for Trema.micrantha}
\item{Trichanthera.gigantea}{Count for Trichanthera.gigantea}
\item{Trichilia.pallida}{Count for Trichilia.pallida}
\item{Trichilia.tuberculata}{Count for Trichilia.tuberculata}
\item{Trichospermum.galeottii}{Count for Trichospermum.galeottii}
\item{Triplaris.cumingiana}{Count for Triplaris.cumingiana}
\item{Trophis.caucana}{Count for Trophis.caucana}
\item{Trophis.racemosa}{Count for Trophis.racemosa}
\item{Turpinia.occidentalis}{Count for Turpinia.occidentalis}
\item{Unonopsis.pittieri}{Count for Unonopsis.pittieri}
\item{Virola.multiflora}{Count for Virola.multiflora}
\item{Virola.sebifera}{Count for Virola.sebifera}
\item{Virola.surinamensis}{Count for Virola.surinamensis}
\item{Vismia.baccifera}{Count for Vismia.baccifera}
\item{Vochysia.ferruginea}{Count for Vochysia.ferruginea}
\item{Xylopia.macrantha}{Count for Xylopia.macrantha}
\item{Zanthoxylum.ekmanii}{Count for Zanthoxylum.ekmanii}
\item{Zanthoxylum.juniperinum}{Count for Zanthoxylum.juniperinum}
\item{Zanthoxylum.panamense}{Count for Zanthoxylum.panamense}
\item{Zanthoxylum.setulosum}{Count for Zanthoxylum.setulosum}
\item{Zuelania.guidonia}{Count for Zuelania.guidonia}}
}
\source{
https://www.science.org/doi/10.1126/science.1066854
}
\usage{
BCI
}
\description{
This dataset contains tree counts from Barro-Colorado Island. It has 50
rows each representing the counts taken from a separate one hectare plot
for each of the 225 species (columns)
}
\keyword{datasets}
