% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diversity_indices.R
\name{calc_diversity_indices}
\alias{calc_diversity_indices}
\title{Calculate biodiversity summary indices from count data}
\usage{
calc_diversity_indices(data)
}
\arguments{
\item{data}{A data frame of count data,
with sites as rows and species as columns.}
}
\value{
A data frame with sites as rows and diversity indices as columns.
The columns are: abundance,
species richness,
Shannon diveristy index,
Simpson diversity index,
Chao1,
Difference between Choa1 and species richness.
}
\description{
Calculate biodiversity summary indices from count data
}
\examples{
ind <- calc_diversity_indices(BCI)
}
