% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_abun_curve.R
\name{create_abun_curve}
\alias{create_abun_curve}
\title{Create species abundance curve}
\usage{
create_abun_curve(species, counts)
}
\arguments{
\item{species}{A list of species}

\item{counts}{Associated counts for each species}
}
\value{
A abundance curve ggplot object
}
\description{
Create species abundance curve
}
\examples{
p <- (species, counts)
}
